restructuring display functions to a separate package
